Move RBi related cops to `rbi/`
