+++ lib/rubocop/cop/sorbet/rbi/forbid_extend_t_sig_helpers_in_shims.rb
