+++ lib/rubocop/cop/sorbet/rbi/single_line_rbi_class_module_definitions.rb
