+++ lib/rubocop/cop/sorbet_cops.rb
@@ -1,13 +1,14 @@
 # frozen_string_literal: true
 require_relative 'sorbet/binding_constants_without_type_alias'
 require_relative 'sorbet/constants_from_strings'
-require_relative 'sorbet/forbid_extend_t_sig_helpers_in_shims'
 require_relative 'sorbet/forbid_superclass_const_literal'
 require_relative 'sorbet/forbid_include_const_literal'
 require_relative 'sorbet/forbid_untyped_struct_props'
-require_relative 'sorbet/single_line_rbi_class_module_definitions'
 require_relative 'sorbet/one_ancestor_per_line'
 require_relative 'sorbet/callback_conditionals_binding'
+
+require_relative 'sorbet/rbi/forbid_extend_t_sig_helpers_in_shims'
+require_relative 'sorbet/rbi/single_line_rbi_class_module_definitions'
 
 require_relative 'sorbet/signatures/allow_incompatible_override'
 require_relative 'sorbet/signatures/checked_true_in_signature'
